// Capstone Project/src/app/pages/admin/admin-type-of-rental-management-page/screen/types.ts
export enum adminPageTypeOfRentalEnum {
  GET_DATA_TYPE_OF_RENTAL_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_TYPE_OF_RENTAL_REQUEST',
  GET_DATA_TYPE_OF_RENTAL_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_TYPE_OF_RENTAL_SUCCESS',
  GET_DATA_TYPE_OF_RENTAL_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_TYPE_OF_RENTAL_FAILURE',

  UPDATE_DATA_TYPE_OF_RENTAL_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_TYPE_OF_RENTAL_REQUEST',
  UPDATE_DATA_TYPE_OF_RENTAL_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_TYPE_OF_RENTAL_SUCCESS',
  UPDATE_DATA_TYPE_OF_RENTAL_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_TYPE_OF_RENTAL_FAILURE',

  CREATE_DATA_TYPE_OF_RENTAL_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_TYPE_OF_RENTAL_REQUEST',
  CREATE_DATA_TYPE_OF_RENTAL_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_TYPE_OF_RENTAL_SUCCESS',
  CREATE_DATA_TYPE_OF_RENTAL_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_TYPE_OF_RENTAL_FAILURE',

  DELETE_DATA_TYPE_OF_RENTAL_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_TYPE_OF_RENTAL_REQUEST',
  DELETE_DATA_TYPE_OF_RENTAL_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_TYPE_OF_RENTAL_SUCCESS',
  DELETE_DATA_TYPE_OF_RENTAL_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_TYPE_OF_RENTAL_FAILURE',

  CLEAR_DATA_TYPE_OF_RENTAL = 'APP/ADMIN_PAGE/CLEAR_DATA_TYPE_OF_RENTAL',
}